Port the code-model to the embedded free-tree system. This means that now inserting/removing declarations from the code-model is much more efficient, especially when the code-model for a filename contains many items. No more iteration of all items. Also the returned item-list is now sorted, so that it's very fast to check whether the code-model of a specific file contains an identifier.

+++ language/duchain/repositories/itemrepository.h
@@ -1178,7 +1178,7 @@ class ItemRepository : public AbstractItemRepository {
               if(totalAvailableSpace > totalSize) {
                 Q_ASSERT(extent);
                 ///We can merge these buckets into one monster-bucket that can hold the data
-                Q_ASSERT(m_freeSpaceBuckets[a-extent] == rangeStart);
+                Q_ASSERT((uint)m_freeSpaceBuckets[a-extent] == (uint)rangeStart);
                 m_freeSpaceBuckets.remove(a-extent, extent+1);
                 m_freeSpaceBucketsSize = m_freeSpaceBuckets.size();
                 useBucket = rangeStart;
